mining schedule hook correctly calls mining start and stop according to schedule
